packages/rapidforge-vault/vault_engine.py
import nacl.secret
import nacl.utils
from argon2 import PasswordHasher
import sys
import json
import base64

# Use Argon2id for KDF
ph = PasswordHasher(
    time_cost=3, 
    memory_cost=65536, 
    parallelism=4, 
    hash_len=32
)

def derive_key(passphrase: str, salt: bytes) -> bytes:
    # Argon2id produces a stable key from passphrase and salt
    # In production, use a secure salt
    return ph.hash(passphrase + salt.hex()).encode()[:32]

def encrypt_data(data: bytes, key: bytes) -> str:
    box = nacl.secret.SecretBox(key)
    nonce = nacl.utils.random(nacl.secret.SecretBox.NONCE_SIZE)
    encrypted = box.encrypt(data, nonce)
    return base64.b64encode(encrypted).decode('utf-8')

def decrypt_data(encrypted_str: str, key: bytes) -> bytes:
    box = nacl.secret.SecretBox(key)
    encrypted = base64.b64decode(encrypted_str)
    return box.decrypt(encrypted)

if __name__ == "__main__":
    # Simple CLI bridge for Node.js
    command = sys.argv[1]
    # ... logic for parsing EHP messages and executing ...
    print(json.dumps({"status": "ready"}))
